Squashed commit of the following:
commit ee7b95695e2ed629412f35e01c760e8291147265
Author: Paul Crashley <paulcrashley@gmail.com>
Date:   Wed Nov 28 14:47:48 2018 +0000

    Updates ttl to float

+++ src/CachableQueryBuilder.php
@@ -22,14 +22,14 @@ class CachableQueryBuilder extends AbstractCachable implements CachableQueryBuil
      *
      * @param QueryBuilderInterface $queryBuilder
      * @param Cache                 $cache
-     * @param int                   $ttl
+     * @param float                 $ttl
      * @param Request|null          $request
      * @param null|string           $env
      */
     public function __construct(
         QueryBuilderInterface $queryBuilder,
         Cache $cache,
-        int $ttl = 10,
+        float $ttl = 10,
         ?Request $request = null,
         ?string $env = null
     ) {
